test(agent): split provider loop coverage
